Rearrange the unit tests
