--- stack/docker-compose.base-with-services.yml
@@ -1,19 +1,0 @@
----
-version: '3.4'
-
-services:
-
-    aiidalab:
-        image: ${REGISTRY:-}${BASE_WITH_SERVICES_IMAGE:-aiidalab/base-with-services}:${VERSION:-newly-build}
-        environment:
-            TZ: Europe/Zurich
-            DOCKER_STACKS_JUPYTER_CMD: notebook
-            SETUP_DEFAULT_AIIDA_PROFILE: 'true'
-            AIIDALAB_DEFAULT_APPS: ''
-        volumes:
-            - aiidalab-home-folder:/home/jovyan
-        ports:
-            - "0.0.0.0:${AIIDALAB_PORT:-}:8888"
-
-volumes:
-    aiidalab-home-folder:
